자바 컨트롤러 및 기능별 패키지 분류, 재정리 210722
