+++ jobdamoa/src/main/java/com/ch/jobdamoa/scrapcheck/ExecScrapPy.java
@@ -7,6 +7,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.scheduling.annotation.Scheduled;
 
+import com.ch.jobdamoa.scrapping.ExecScrapping;
+
 public class ExecScrapPy {
 	
 	private static final Logger logger = LoggerFactory.getLogger(ExecScrapPy.class);
@@ -16,7 +18,9 @@ public class ExecScrapPy {
 		 Calendar calendar = Calendar.getInstance();
 		 SimpleDateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
 		 logger.info("스케줄 실행 : " + dateFormat.format(calendar.getTime()));
-		 Scrapping sc = new Scrapping();
-		 sc.doProgrammers();
+		 
+		 // 5분마다 주기적으로 실행하는 프로세스
+		 ExecScrapping es = new ExecScrapping();
+		 es.doProgrammers();
 	}
 }
