+++ jobdamoa/src/main/java/com/ch/jobdamoa/scrapping/Scrapping.java
@@ -1,0 +1,7 @@
+package com.ch.jobdamoa.scrapping;
+
+public interface Scrapping {
+
+	void doProgrammers();
+
+}
