+++ jobdamoa/src/main/java/com/ch/jobdamoa/service/ScrapService.java
@@ -1,5 +1,11 @@
 package com.ch.jobdamoa.service;
 
+import java.util.List;
+
+import com.ch.jobdamoa.model.ScrapJson;
+
 public interface ScrapService {
 
+	List<ScrapJson> scrapProgrammers();
+	
 }
